Move config to pkg directory
This would reorganize the code in a more standarized way: https://github.com/golang-standards/project-layout
